(REF) APIv4 FieldSpec - Extract SqlSpecTrait (tableName, columnName, operators, sqlRenderer, sqlFilters)
